validade-inteligente-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
openai==0.28.1
requests==2.31.0
Werkzeug==2.3.7
gunicorn==21.2.0
celery==5.3.1
bcrypt==4.0.1
email-validator==2.0.0
Pillow==10.0.0
python-multipart==0.0.6
mercadopago==2.2.1